Move echo server-specific files to echo server directory
