+++ network/components/lwip_timer.c
@@ -1,4 +1,4 @@
-#include <sddf/util/timer.h>
+#include "timer.h"
 
 #define TIMER_CH 1 
 #define LWIP_TICK_MS 100
